Split check urls & check
