+++ .travis.yml
@@ -9,3 +9,5 @@ addons:
     - imagemagick
 script:
 - python check.py
+- python check_urls.py
+
