+++ README.rst
@@ -42,6 +42,7 @@ Check appliance files
 .. code:: bash
 
     python3 check.py
+    python3 check_urls.py
 
 You need to install `imagemagick` before running check.py.
