+++ check.py
@@ -19,31 +19,7 @@ import os
 import jsonschema
 import json
 import sys
-import socket
 import subprocess
-import urllib.request
-from multiprocessing import Pool
-
-
-class MyHTTPRedirectHandler(urllib.request.HTTPRedirectHandler):
-    def redirect_request(self, req, fp, code, msg, hdrs, newurl):
-        return None
-
-urllib.request.install_opener(urllib.request.build_opener(MyHTTPRedirectHandler))
-
-def check_url(args):
-    url, appliance = args
-    try:
-        print("Check " + url)
-        req = urllib.request.Request(url, method='HEAD')
-        urllib.request.urlopen(req, timeout=20)
-    except urllib.error.HTTPError as err:
-        if err.getcode() >= 400:
-            raise Exception('Error with url ' + url + ' - ' + str(err))
-    except urllib.error.URLError as err:
-        raise Exception('Invalid URL ' + url)
-    except socket.timeout as e:
-        raise Exception('Timeout URL ' + url)
 
 
 def check_appliance(appliance):
@@ -81,26 +57,6 @@ def check_appliance(appliance):
                 sys.exit(1)
 
 
-def check_urls(pool, appliance):
-    with open(os.path.join('appliances', appliance)) as f:
-        appliance_json = json.load(f)
-
-    calls = []
-
-    for image in appliance_json['images']:
-        if 'direct_download_url' in image:
-            calls.append((image['direct_download_url'], appliance))
-        if 'download_url' in image:
-            calls.append((image['download_url'], appliance))
-
-    if 'vendor_url' in appliance_json:
-        calls.append((appliance_json['vendor_url'], appliance))
-    if 'documentation_url' in appliance_json:
-        calls.append((appliance_json['documentation_url'], appliance))
-    if 'product_url' in appliance_json:
-        calls.append((appliance_json['product_url'], appliance))
-    return calls
-
 def check_packer(packer):
     path = os.path.join('packer', packer)
     if not os.path.isdir(path):
@@ -124,14 +80,10 @@ def check_symbol(symbol):
 
 
 def main():
-    pool = Pool(processes=8)
-
-    calls_check_url = []
     print("=> Check appliances")
     for appliance in os.listdir('appliances'):
         print('Check {}'.format(appliance))
         check_appliance(appliance)
-        calls_check_url += check_urls(pool, appliance)
     print("=> Check symbols")
     for symbol in os.listdir('symbols'):
         if symbol.endswith('.svg'):
@@ -140,14 +92,6 @@ def main():
     print("=> Check packer files")
     for packer in os.listdir('packer'):
         check_packer(packer)
-    print("=> Check URL in appliances")
-    try:
-        pool.map_async(check_url, calls_check_url).get()
-    except Exception as e:
-        print(e)
-        sys.exit(1)
-    pool.close()
-    pool.join()
     print("Everything is ok!")
 
 if __name__ == '__main__':
